requirements.txt
matplotlib==3.8.4
numpy==2.1.1
pandas==2.2.2
pynwb==2.8.1
PyYAML==6.0.1
PyYAML==6.0.2
scipy==1.14.1
seaborn==0.13.2
